Add `--project` to run tests with all mm/oot asm
This does not check in the results of the tests, I think there needs to
be some more logistics to work out how best to incorporate these tests
into the workflow.

Running the tests with context is *very* slow.

For the project tests, if the decomp crashes, it includes the error
message in the output file and keeps going with the functions it *can*
handle. It also prefixes the output file with the `CRASHED` line.