Unify CType & Type
This change makes all CTypes representable by a Type. This allows us to
track type information for global symbols correctly.

There are some small changes in the test outputs, they fall into the
following categories:

- Pointers to primitive typedefs are no longer preserved
    - Ex: `(int *)` is now uniformly `(s32 *)`
- Global symbol types can now usually be inferred, so there are fewer
type casts for these symbols.
- The format of the `--dump-typemap` was changed to match how variables
are declared. This was a side effect of fixing how the types of function
pointers are printed
- There are now a few more `(u8)` and `(u16)` explicit casts.