Unwrap before comparison in fold_mul_chains
This results in being able to fold more arithmetic chains that were
previously blocked by having EvalOnceExprs in the tree.

This change should be safe because we're already willing to unwrap the
EvalOnceExpr if it could participate in folding.

Spot checked against MM & OOT code. Results in better folding and fewer
temps.