Fix structs getting typed as s32
`get_size_align_bytes` has a super-sketchy fallback to 32 bits when no
size is known. When calling `deref` with an unknown pointer x, offset y
and a 32-bit width, it used to hit this fallback and deduce that the
access was an array access at offset y // 4, and set the type of x to
be a pointer to the dereferenced type. Fix this by using `get_size_bits`
instead (via a new helper `get_size_bytes`). The `sub_align` check
probably wasn't necessary anyway.