Rewrite `TypeData.get_representative()` to avoid recursion
There were two tests from MM that hit the recursion limit: an
initializer array with 4000 elements would create a linked list of 4000
TypeDatas in the UF datastructure.

It might be possible to trigger compaction earlier? For now, its easy
enough to avoid recursion alltogether, and it made it easier to add a
check for cycles.

Example from MM: `code_0x80182CE0`

$ ./run_tests.py --project ../mm --filter=code_0x80182CE0