Use a struct to represent the stack
The struct can optionally be provided as part of the context.

This changeset provides most of the groundwork for allowing structs on
the stack, and for detecting the types of unknown regions of structs
provided in the context.

Added the `exact` option to `.get_deref_field()` to make it slightly
more ergonomic.