Allow multiple `--context` files; add caching/`--no-cache` (#152)
* Allow multiple `--context` files; add `--cache`

Parse multiple `--context` files in sequence. Had to add a slight hack
with `CParser()` to preserve a small bit of state, which is stored in
`TypeMap`.

Add caching, where `ctx.c` is cached in a pickle file at `ctx.c.m2c`. A
hash of the inputs to `build_typemap` is stored in the `TypeMap` itself
and used to validate that the cache is still valid.

Loading a huge context, like from MM, takes about 6 seconds on cache
miss, and about 250ms on cache hit. (Most of that time seems to be in
unpickling; not file I/O or computing the hash.)

Caching might seem like surprising behavior, so I left it disabled by
default. (Maybe it should be enabled by default, and disabled with
`--no-cache`?)

* Turn on caching by default

* `build_typemap()` takes a list of contexts; add test

* Support multiple `--context` flags in tests

* Remove `TypeMap.empty()`

* Set a default for `TypeMap.source_hash`