Bugs; Decomp'd symbols have function type; project test args
The list of symbols that are going to be decomp'd
(`GlobalInfo.local_functions`) are known to have some kind of function
type, even if they do not have type information available in the
`TypeMap`.
This fixes the `AssertionError: fn_type is known to be a function`
errors that a few complicated objects were hitting in the full-project
tests.

Updated `run_tests.py` to set per-project flags, rather than just
configuring the compiler.

Fixed bug in "overlap" warning calculation.

Fixed bug with `replace_first_arg`; take the name/type of the first arg
from the original arg.