Rename `--no-struct-inference` to `--no-unk-inference`; use `is_unk_type()` on GlobalSymbol types
Sorry this broke the `--no-struct-inference` that was added a few hours
ago; but the new name covers the additional use cases.

This lets `mips_to_c` treat global symbols like `UNK_TYPE foo;` from the
context as having an indeterminate type. It's inferred type will be
included in the globals section, if enabled.

This fixes a handful of extraneous casts in the MM project (for example).

Diff for OOT/MM:
https://gist.github.com/zbanks/f030b71d08f5092348decf9274b9b597