Treat GlobalSymbols declared as zero-length arrays as VLAs
If a GlobalSymbol is declared as a zero-length array in the context,
e.g. `int foo[0];`, then it's likely that this represents a
variable-legnth array (VLA) and not that the symbol truly has zero size.

This change mostly affects Paper Mario, as now mips2c will be able to
identify & size script arrays (which have type `s32 [0]`)