Don't restore regs after a fn call w/ a stack ref (#195)
The `local_var_writes` dict is used to track the last value written to
each location on the stack, and is used to elide register restore
statements (i.e. after function calls).

However, if a function is called with a reference to the stack, the
function *may* modify the stack and invalidate the value held in this
dict.

If this is detected, then clear out the `local_var_writes` dict to
disable the register-restore logic.