Match stack args to their slots based on offset
Use the expected `offset` of each function arg to choose the correct arg
from `subroutine_args`, and emit an error if it is missing.

Additionally, if the function has `params_known`, it is not nessassary
to re-unify the parameter types. This was causing frame shift issues when
there was a return struct or a >4-byte parameter.

Diffs:

https://gist.github.com/zbanks/824e7811836821cca181ed675bbac4c0