Allow matching on literal relations in asm patterns
The goal of this is to eventually aid a double-to-int PPC pattern, which
involve two consecutive stack stores. The check for the stores being
consecutive cannot easily be made in the function that verifies the match/
computes the replacement, because MWCC can reorder asm instructions within
patterns with surrounding instructions, and when adding support for reorderings
to the pattern matching code we'll want verification checks to be incremental
and done as part of the pattern matching.