Add scripts for disassembling PPC ELF files (WIP)
The goal of these scripts is to add PPC end-to-end tests. They still
need to be incorporated into `add_test.py`.

The `ppc_disasm.py` script is modeled after `doldisasm.py` and produces
similarly-formatted output.

The `elf_file.py` abstraction is mostly based on the code in
`asm_differ`. It could likely be re-used for a MIPS disassembler, if we
add one.