Detect `x|x`; comment about limitions of demangling
`or. $rD, $rS, $rS` is used on PPC as a way to set the condition bits
based on the value of `$rD` (while moving the value into `$rS`).

Added additional comments to `types.py` in `demangle_type` that add
caveats to the argument detection heuristics based on the mangled symbol
name.
Most of these situations are still relatively rare; so it's still worth
keeping the `this`-insertion (imo).