Add `split_arg_list` to split quoted, comma-sep lists
This uses the `csv` module from the stdlib. This was chosen for
simplicity, but I did some microbenchmarks on my laptop to ensure
it didn't have unexpectedly poor performance.

The implementation of `split_arg_list` is about 3x slower than
`[x.strip() for x in args.split(",")]`, which doesn't handle quoting.

However, it is about 3x faster than a simple char-by-char parser that
would handle quoting/escaping (written in the style of `parse_arg_elems`)

(But again, I don't think this is a big performance issue either way)