Implement PPC carry/extended ops, floating intriniscs, clz (#207)
* Implement PPC carry/extended ops, floating intriniscs, clz

This rounds out the most popular instructions.
Added instructions:
- `addc`
- `adde`
- `addze`
- `subfc`
- `subfe`
- `subfze`
- `andc`
- `eqv`
- `cntlzw`
- `fres`
- `frsqrte`
- `fsel`

The carry/extended behavior of the `-c`/`-e` instructions is not
implemented. Usually these instructions are used to implement 64-bit
integer arithmetic, but the codebase isn't currently equipped to easily
handle this. (The instructions may be interleaved, and the value may be
in an arbitrary pair of registers.)

Added a pattern to `fold_divmod` to rewrite `CLZ(x) >> 5` into `x == 0`.
This accounts for ~80% of `cntlzw` usage (and is why its implemented as
a `UnaryOp` instead of a `fn_op`).

Also expanded the pattern for modulo-by-power-of-2 to match the MWCC
version, which emits `(x - (x >> N) * M)` instead of `(x - (x / N) * N)`.

The other comparison patterns aren't detected yet (see the `comparison`
e2e test).

* Rework `handle_rlwinm`; add `CarryBit`

- Use `CarryBit` (which shows up as `MIPS2C_CARRY`) on extended
instructions
- Replace `MIPS2C_FRES`/`MIPS2C_FRSQRTE` with `__fres`/`__frsqrte`
intrinsics
- Modify `fold_divmod` to detect `+ MIPS2C_CARRY`
- Move `CLZ(x) >> 5` code into `replace_clz_shift`, and add a case for
`(a - b) == 0` --> `a == b`
- Avoid performing any folding/simplification on rlw terms with an `|`

* Remove special case casts from handle_rlwinm

* PR feedback (casts, fold_divmod)

* Lower bits of rlwinm are u32 not int