Expand % pattern to expressions
The previous pattern for detecting mod operations was limited to taking
the mod of a `Literal` constant.

However, PPC has no mod instruction, so mod is implemented using the
same series of instructions that GCC uses to mod by a constant in MIPS,
`x % y` becomes `(x - ((x / y) * y))`.

This change produces no diff in any of the MIPS project corpuses, but
makes several improvements to the PPC ones.