Refactor translate_node_body() into NodeState class
This change "mechanically" moves state out of local vars & functions in
`translate_node_body()` and into a dedicated `NodeState` class.

This makes it more clear how to refactor the `process_instruction()` to
be arch-specific (and move it out of translate.py).

The naming is a bit clumsy for now, and there wasn't much effort put
in into reducing the scope of `NodeState`.