Add `--visualize=asm` option for SVG graph w/ asm
Make `--visualize` take an optional argument, either `asm` or `c`, and
default to the existing behavior of `c`.

The new `--visualize=asm` graph prints the post-processed asm in the
flow graph, instead of the translated C source. I had been manually
editing `flow_graph.py` to make these graphs, so I figured it was worth
including.

I could imagine adding a 3rd option in the future to output a flowgraph
without performing any of the pattern substitutions (maybe `raw_asm`?)
-- but I think that'd be a bit trickier to add, and I don't have any
use-case in mind for that.

I did not update the website to make this an option, I think this is
more useful as a developer/debugging tool than for actual decomp?