Handle values saved across function calls in prevent_later_uses
This doesn't actually currently have an effect, because we always
generate temps for reads/function call results (the things that calls
clobber) as a side effect of explicitly emitting the spill statement.

However, if in the future we start eliding spills across function calls
this will be needed.

Skipping temps for global reads, as done in the next commit after this,
also makes this necessary.